ultimos cambios

+++ student/models.py
@@ -85,7 +85,7 @@ class EntregaArchivos(models.Model):
                                          verbose_name=('Codigo Estudiantil'), unique=True)
     formato_guia = models.FileField(upload_to="archivos/", null=True, blank=True,
                                     verbose_name=('Plantilla Guia'))
-    nota = models.IntegerField(max_length=10, null=True, blank=False, verbose_name=('Ingrese Nota'))
+    nota = models.FloatField(max_length=10, null=True, blank=False, verbose_name=('Ingrese Nota'))
     observaciones = models.TextField(null=True, blank=False, verbose_name=('Ingrese Observaciones'))
     rubrica = models.FileField(upload_to="archivos/", null=True, blank=True,
                                verbose_name=('Rebrica de calificacion'))
@@ -104,7 +104,7 @@ class EntregaArchivosModulo3(models.Model):
                                          verbose_name=('Codigo Estudiantil'), unique=True)
     formato_guia = models.FileField(upload_to="archivos/", null=True, blank=True,
                                     verbose_name=('Plantilla Guia'))
-    nota = models.IntegerField(max_length=10, null=True, blank=False, verbose_name=('Ingrese Nota'))
+    nota = models.FloatField(max_length=10, null=True, blank=False, verbose_name=('Ingrese Nota'))
     observaciones = models.TextField(null=True, blank=False, verbose_name=('Ingrese Observaciones'))
     rubrica = models.FileField(upload_to="archivos/", null=True, blank=True,
                                verbose_name=('Rebrica de calificacion'))
